Make list command an abstract class
