+++ src/main/java/seedu/address/logic/parser/ListCommandParser.java
@@ -6,6 +6,7 @@ import static seedu.address.logic.parser.AddressBookParser.BASIC_COMMAND_FORMAT;
 
 import java.util.regex.Matcher;
 
+import seedu.address.logic.commands.ListCaseCommand;
 import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
@@ -17,6 +18,7 @@ public class ListCommandParser implements Parser<ListCommand> {
      * @throws ParseException if the user input does not conform the expected format
      */
     public ListCommand parse(String args) throws ParseException {
+
         final Matcher matcher = BASIC_COMMAND_FORMAT.matcher(args.trim());
         if (!matcher.matches()) {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, ListCommand.MESSAGE_USAGE));
@@ -24,11 +26,13 @@ public class ListCommandParser implements Parser<ListCommand> {
 
         final String commandWord = matcher.group("commandWord");
 
+        //TODO: check state of the program here. If main page, check if TYPE_CASE. If other page, check other TYPES.
+
         if (!commandWord.equals(TYPE_CASE)) {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, ListCommand.MESSAGE_USAGE));
         }
 
-        return new ListCommand();
+        return new ListCaseCommand();
 
     }
 }
