+++ src/test/java/seedu/address/logic/parser/ListCommandParserTest.java
@@ -6,6 +6,7 @@ import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSucces
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.ListCaseCommand;
 import seedu.address.logic.commands.ListCommand;
 
 public class ListCommandParserTest {
@@ -13,8 +14,8 @@ public class ListCommandParserTest {
     private ListCommandParser parser = new ListCommandParser();
 
     @Test
-    public void parse_validArgs_returnsListCommand() {
-        assertParseSuccess(parser, "case", new ListCommand());
+    public void parse_validArgs_returnsListCaseCommand() {
+        assertParseSuccess(parser, "case", new ListCaseCommand());
     }
 
     @Test
